global-setup: extend ingestion timeout to 240s
During initial setup of data ingestion of SBOMs
can take longer than current 60s timeout allows.

Timeout should be only safety measure here
(its not a perf test, only one time pre-test setup)
so increase time limit to double
(in case of UI: 2min for SBOMs and 2min for Advisories).

As with addition of more files we may need to increase
this more in future, using constant defined for it
at same place as the list of files to be ingested.