global-setup: extend ingestion timeout to 240s (#24)
During initial setup, ingestion of SBOMs
can take longer than current 60s timeout allows.

Timeout should be only safety measure here
(its not a perf test, only one time pre-test setup)
so increase time limit to double
(in case of UI: 2min for SBOMs and 2min for Advisories).

As with addition of more files we may need to increase
this again in future, switching to use constant defined for it
at same place as the list of files to be ingested.