UI: Added Vulnerability Explorer tests
Due to some changes in the requirements that were not reflected in the
feature file, this turned out to be actually somewhat simpler than
anticipated. Creating this MR to keep it easier to track in relation to
the respective Jira.

---------

Co-authored-by: Queria Sa-Tas <queria@sa-tas.net>
(cherry picked from commit 071eef5e5dde06a90b91c5461d2109a51b4daa07)